[refactor] 오늘의 옷장 카드 layout 수정 & slide 추가
