+++ package.json
@@ -20,6 +20,7 @@
     "react-router-dom": "^6.8.2",
     "react-scripts": "5.0.1",
     "styled-components": "^5.3.8",
+    "swiper": "^9.1.1",
     "typescript": "^4.9.5",
     "web-vitals": "^2.1.4"
   },
