+++ src/page/Home.tsx
@@ -1,112 +1,14 @@
 import React, { useState } from 'react';
 import styled from 'styled-components';
-import { ReactComponent as BookMarkSvg } from '../styles/images/icons/bookmark.svg';
-import { ReactComponent as LikesSvg } from '../styles/images/icons/likes.svg';
-import { ReactComponent as FollowSvg } from '../styles/images/icons/follow.svg';
+import { Swiper, SwiperSlide } from 'swiper/react';
 import TodayCardModal from 'components/Home/TodayCardModal';
+import TodayCard from 'components/Home/TodayCard';
+import 'swiper/css';
+import 'swiper/css/navigation';
 
 const Card = styled.section`
   margin-top: 4px;
-  background-color: ${({ theme }) => theme.colors.gray3};
-  border-radius: ${({ theme }) => theme.borderRadius.borderRadius30};
-  height: 200px;
-  padding: 23px 0;
 `;
-
-const CardHeader = styled.div`
-  display: flex;
-  justify-content: space-between;
-  padding: 0 16px;
-`;
-
-const CardHeaderUser = styled.div`
-  display: flex;
-  align-items: center;
-`;
-
-const CardHeaderProfile = styled.div`
-  width: 40px;
-  height: 40px;
-  border-radius: ${({ theme }) => theme.borderRadius.borderRadius50};
-  border: 1px solid ${({ theme }) => theme.colors.gray7};
-  background-color: ${({ theme }) => theme.colors.gray9};
-`;
-
-const CardHeaderInfo = styled.div`
-  margin-left: 8px;
-  display: flex;
-  flex-direction: column;
-
-  & strong {
-    font-size: 17px;
-    font-weight: 600;
-  }
-
-  & span {
-    font-size: 13px;
-    font-weight: 500;
-    margin-top: 2px;
-  }
-`;
-
-const CardHeaderUtil = styled.ul`
-  display: flex;
-  & li {
-    width: 40px;
-    height: 40px;
-    background-color: ${({ theme }) => theme.colors.gray4};
-    border-radius: ${({ theme }) => theme.borderRadius.borderRadius50};
-
-    & button {
-      width: 100%;
-      height: 100%;
-      display: flex;
-      justify-content: center;
-      align-items: center;
-
-      & svg {
-        & path {
-          fill: ${({ theme }) => theme.colors.gray6};
-        }
-      }
-    }
-  }
-
-  & li + li {
-    margin-left: 8px;
-  }
-`;
-
-const CardItems = styled.ul`
-  margin: 24px 0 0 16px;
-  display: flex;
-  overflow-x: scroll;
-  &::-webkit-scrollbar {
-    display: none;
-  }
-
-  & li + li {
-    margin-left: 12px;
-  }
-
-  & li {
-    & span {
-      font-size: 12px;
-      font-weight: 700;
-      color: ${({ theme }) => theme.colors.gray7};
-    }
-
-    & div {
-      margin-top: 5px;
-      width: 72px;
-      height: 72px;
-      background-color: ${({ theme }) => theme.colors.gray5};
-      border-radius: ${({ theme }) => theme.borderRadius.borderRadius20};
-    }
-  }
-`;
-
-const CardItem = styled.li``;
 
 const Look = styled.section`
   & > h2 {
@@ -131,83 +33,39 @@ const LookTabSwiper = styled.div`
   }
 `;
 
+const CardSlide = styled.div`
+  padding: 23px 0;
+  background-color: ${({ theme }) => theme.colors.gray3};
+  border-radius: ${({ theme }) => theme.borderRadius.borderRadius30};
+`;
+
 function Home() {
   const [isOpen, setIsOpen] = useState<boolean>(false);
-
-  const onClickOpenModal = () => {
-    setIsOpen(true);
-  };
 
   return (
     <>
       <TodayCardModal isOpen={isOpen} setIsOpen={setIsOpen} />
-
-      <Card>
-        <div>
-          <div>
+      <Swiper spaceBetween={20}>
+        <SwiperSlide>
+          <Card>
             <div>
-              <CardHeader>
-                <CardHeaderUser>
-                  <CardHeaderProfile></CardHeaderProfile>
-                  <CardHeaderInfo>
-                    <strong>Anonymous</strong>
-                    <span>11월 3일 옷장</span>
-                  </CardHeaderInfo>
-                </CardHeaderUser>
-                <CardHeaderUtil>
-                  <li>
-                    <button>
-                      <LikesSvg />
-                    </button>
-                  </li>
-                  <li>
-                    <button>
-                      <BookMarkSvg />
-                    </button>
-                  </li>
-                  <li>
-                    <button>
-                      <FollowSvg />
-                    </button>
-                  </li>
-                </CardHeaderUtil>
-              </CardHeader>
-              <CardItems>
-                <CardItem>
-                  <button type="button" onClick={onClickOpenModal}>
-                    <span>outer</span>
-                    <div></div>
-                  </button>
-                </CardItem>
-                <CardItem>
-                  <button type="button" onClick={onClickOpenModal}>
-                    <span>outer</span>
-                    <div></div>
-                  </button>
-                </CardItem>
-                <CardItem>
-                  <button type="button" onClick={onClickOpenModal}>
-                    <span>outer</span>
-                    <div></div>
-                  </button>
-                </CardItem>
-                <CardItem>
-                  <button type="button" onClick={onClickOpenModal}>
-                    <span>outer</span>
-                    <div></div>
-                  </button>
-                </CardItem>
-                <CardItem>
-                  <button type="button" onClick={onClickOpenModal}>
-                    <span>outer</span>
-                    <div></div>
-                  </button>
-                </CardItem>
-              </CardItems>
+              <CardSlide>
+                <TodayCard setIsOpen={setIsOpen} />
+              </CardSlide>
             </div>
-          </div>
-        </div>
-      </Card>
+          </Card>
+        </SwiperSlide>
+        <SwiperSlide>
+          <Card>
+            <div>
+              <CardSlide>
+                <TodayCard setIsOpen={setIsOpen} />
+              </CardSlide>
+            </div>
+          </Card>
+        </SwiperSlide>
+      </Swiper>
+
       <Look>
         <h2>오늘의 추천 look</h2>
