+++ src/types/Home/index.ts
@@ -1,4 +1,8 @@
-export interface ITodayCardModalProps {
+export interface ITodayCardModalProps extends ITodayCardProps {
   isOpen: boolean;
+  //setIsOpen: React.Dispatch<React.SetStateAction<boolean>>;
+}
+
+export interface ITodayCardProps {
   setIsOpen: React.Dispatch<React.SetStateAction<boolean>>;
 }
